# lib/api-zod/src/generated/types/discordUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  globalName?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  banner?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  accentColor?: number | null;
  bot: boolean;
  publicFlags: number;
  createdAt: string;
  /** @nullable */
  clanTag?: string | null;
  /** @nullable */
  clanBadgeHash?: string | null;
  /** @nullable */
  clanGuildId?: string | null;
  premiumType?: number;
}